decompose+rename+flag fixes

+++ src/s21_dectostr.c
@@ -9,7 +9,7 @@ char* dectostr(const s21_decimal dec) {
     str = calculate_decimal(str, dec.bits);
     
     if (negative || e) {
-        if (strlen(str) <= e) str = add_width(str, e - strlen(str) + 1, '0', true);
+        if ((int)strlen(str) <= e) str = add_width(str, e - strlen(str) + 1, '0', true);
         int i = strlen(str) + negative + (bool)e;
         str = (char*)realloc(str, (i + 1) * sizeof(char));
